add new AudioTags class
Yay! I can set file metadata tags without bothering what file type it is (or if it's supported). Another great thing is, this inherits from a dict, so it acts just like a dict, but with makes every key a tag name I specified in the `_tags.py` file.

I named `_tags.py` with an underscore because I couldn't think of a better name, and I wanted to use `tags` as a variable name in my code. This file also has the ability to be expandable, so I can add different file types (like ogg).